Read from file/cache if exists, otherwise save to file/cache.

+++ GetAndSave.py
@@ -24,7 +24,6 @@ elif not os.path.isdir(d):
 
 for url in urls.values():
     url_header = {}
-    # fn = hash(url)  # TODO: Inconsistent output? Investigate.
     url_md5 = md5(str(url).encode('utf-8')).hexdigest()
     header_cache_filename = d + '/' + url_md5 + '.header'  # TODO: Refractor path management
     content_cache_filename = d + '/' + url_md5 + '.content'
@@ -42,10 +41,12 @@ for url in urls.values():
     else:
         url_header['url'] = response.geturl()  # In case of redirects
         content = response.read()
-        # Decode from bytes immediately? Or do that after the request/read caching process?
-        # if url_header['Content-Type'] == 'text/html':
-        #     content = content.decode()  # Won't work with 'wb' writes.
-        with open(content_cache_filename, 'wb') as content_cache_file:
-            content_cache_file.write(content)
+        if os.path.isfile(content_cache_filename):
+            with open(content_cache_filename, 'rb') as content_cache_file:
+                print(content_cache_file.read())
+        else:
+            with open(content_cache_filename, 'wb') as content_cache_file:
+                content_cache_file.write(content)
+            print('Saved to cache.')
     with open(header_cache_filename, 'w') as header_cache_file:
         json.dump(url_header, header_cache_file)
